+++ aicontent-20240611/alibabacloud_aicontent20240611/models/_conversation_dto.py
@@ -20,6 +20,7 @@ class ConversationDTO(DaraModel):
         self.delete_tag = delete_tag
         self.gmt_create = gmt_create
         self.gmt_modified = gmt_modified
+        # ID
         self.id = id
         self.message_count = message_count
         self.model_ids = model_ids