+++ aicontent-20240611/alibabacloud_aicontent20240611/models/_model_router_query_client_list_response_body.py
@@ -10,50 +10,31 @@ from darabonba.model import DaraModel
 class ModelRouterQueryClientListResponseBody(DaraModel):
     def __init__(
         self,
-        data: List[main_models.ClientDTO] = None,
+        data: main_models.ModelRouterQueryClientListResponseBodyData = None,
         err_code: str = None,
         err_message: str = None,
         http_status_code: int = None,
-        max_results: int = None,
-        next_token: str = None,
-        page_index: int = None,
-        page_size: int = None,
         request_id: str = None,
-        skip: int = None,
         success: bool = None,
-        total_count: int = None,
     ):
         self.data = data
         self.err_code = err_code
         self.err_message = err_message
         self.http_status_code = http_status_code
-        # maxResults
-        self.max_results = max_results
-        # nextToken
-        self.next_token = next_token
-        self.page_index = page_index
-        self.page_size = page_size
         self.request_id = request_id
-        # skip
-        self.skip = skip
         self.success = success
-        self.total_count = total_count
 
     def validate(self):
         if self.data:
-            for v1 in self.data:
-                 if v1:
-                    v1.validate()
+            self.data.validate()
 
     def to_map(self):
         result = dict()
         _map = super().to_map()
         if _map is not None:
             result = _map
-        result['data'] = []
         if self.data is not None:
-            for k1 in self.data:
-                result['data'].append(k1.to_map() if k1 else None)
+            result['data'] = self.data.to_map()
 
         if self.err_code is not None:
             result['errCode'] = self.err_code
@@ -64,39 +45,19 @@ class ModelRouterQueryClientListResponseBody(DaraModel):
         if self.http_status_code is not None:
             result['httpStatusCode'] = self.http_status_code
 
-        if self.max_results is not None:
-            result['maxResults'] = self.max_results
-
-        if self.next_token is not None:
-            result['nextToken'] = self.next_token
-
-        if self.page_index is not None:
-            result['pageIndex'] = self.page_index
-
-        if self.page_size is not None:
-            result['pageSize'] = self.page_size
-
         if self.request_id is not None:
             result['requestId'] = self.request_id
 
-        if self.skip is not None:
-            result['skip'] = self.skip
-
         if self.success is not None:
             result['success'] = self.success
-
-        if self.total_count is not None:
-            result['totalCount'] = self.total_count
 
         return result
 
     def from_map(self, m: dict = None):
         m = m or dict()
-        self.data = []
         if m.get('data') is not None:
-            for k1 in m.get('data'):
-                temp_model = main_models.ClientDTO()
-                self.data.append(temp_model.from_map(k1))
+            temp_model = main_models.ModelRouterQueryClientListResponseBodyData()
+            self.data = temp_model.from_map(m.get('data'))
 
         if m.get('errCode') is not None:
             self.err_code = m.get('errCode')
@@ -107,29 +68,86 @@ class ModelRouterQueryClientListResponseBody(DaraModel):
         if m.get('httpStatusCode') is not None:
             self.http_status_code = m.get('httpStatusCode')
 
+        if m.get('requestId') is not None:
+            self.request_id = m.get('requestId')
+
+        if m.get('success') is not None:
+            self.success = m.get('success')
+
+        return self
+
+class ModelRouterQueryClientListResponseBodyData(DaraModel):
+    def __init__(
+        self,
+        list: List[main_models.ClientDTO] = None,
+        max_results: str = None,
+        next_token: str = None,
+        page: int = None,
+        page_size: int = None,
+        total: int = None,
+    ):
+        self.list = list
+        self.max_results = max_results
+        self.next_token = next_token
+        self.page = page
+        self.page_size = page_size
+        self.total = total
+
+    def validate(self):
+        if self.list:
+            for v1 in self.list:
+                 if v1:
+                    v1.validate()
+
+    def to_map(self):
+        result = dict()
+        _map = super().to_map()
+        if _map is not None:
+            result = _map
+        result['list'] = []
+        if self.list is not None:
+            for k1 in self.list:
+                result['list'].append(k1.to_map() if k1 else None)
+
+        if self.max_results is not None:
+            result['maxResults'] = self.max_results
+
+        if self.next_token is not None:
+            result['nextToken'] = self.next_token
+
+        if self.page is not None:
+            result['page'] = self.page
+
+        if self.page_size is not None:
+            result['pageSize'] = self.page_size
+
+        if self.total is not None:
+            result['total'] = self.total
+
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        self.list = []
+        if m.get('list') is not None:
+            for k1 in m.get('list'):
+                temp_model = main_models.ClientDTO()
+                self.list.append(temp_model.from_map(k1))
+
         if m.get('maxResults') is not None:
             self.max_results = m.get('maxResults')
 
         if m.get('nextToken') is not None:
             self.next_token = m.get('nextToken')
 
-        if m.get('pageIndex') is not None:
-            self.page_index = m.get('pageIndex')
+        if m.get('page') is not None:
+            self.page = m.get('page')
 
         if m.get('pageSize') is not None:
             self.page_size = m.get('pageSize')
 
-        if m.get('requestId') is not None:
-            self.request_id = m.get('requestId')
-
-        if m.get('skip') is not None:
-            self.skip = m.get('skip')
-
-        if m.get('success') is not None:
-            self.success = m.get('success')
-
-        if m.get('totalCount') is not None:
-            self.total_count = m.get('totalCount')
+        if m.get('total') is not None:
+            self.total = m.get('total')
 
         return self