+++ aicontent-20240611/alibabacloud_aicontent20240611/models/_open_api_multi_response.py
@@ -189,8 +189,6 @@ class OpenApiMultiResponseData(DaraModel):
 
         return self
 
-
-
 class OpenApiMultiResponseDataInferenceJobList(DaraModel):
     def __init__(
         self,